+++ scancodeio/static/main.js
@@ -93,6 +93,42 @@ function setupTabs() {
   });
 }
 
+// Menu
+
+function setupMenu() {
+  const $menuLinks = getAll('.menu a:not(.is-stateless)');
+
+  function activateMenuItem($menuItem) {
+    const activeLink = document.querySelector('.menu .is-active');
+    activeLink.classList.remove('is-active');
+    $menuItem.classList.add('is-active');
+  }
+
+  $menuLinks.forEach(function ($el) {
+    $el.addEventListener('click', function () {
+      activateMenuItem($el)
+    });
+  });
+}
+
+// Form
+
+// Dynamic size for the textarea
+function setupTextarea() {
+  const $dynamicTextareas = getAll('textarea.is-dynamic');
+
+  function setHeight($el) {
+    $el.style.height = "";
+    $el.style.height = $el.scrollHeight + 3 + "px";
+  }
+
+  $dynamicTextareas.forEach(function ($el) {
+    $el.oninput = () => { setHeight($el); }
+    $el.onfocus = () => { setHeight($el); }
+  });
+}
+
+
 // Utils, available globally
 
 function getAll(selector) {
@@ -158,6 +194,8 @@ document.addEventListener('DOMContentLoaded', function () {
   setupOpenModalButtons();
   setupCloseModalButtons();
   setupTabs();
+  setupMenu();
+  setupTextarea();
 
   // Close modals and dropdowns on pressing "escape" key
   document.addEventListener('keydown', function (event) {