+++ scanpipe/api/serializers.py
@@ -129,6 +129,8 @@ class ProjectSerializer(
             "webhook_url",
             "created_date",
             "is_archived",
+            "notes",
+            "settings",
             "pipeline",
             "execute_now",
             "input_sources",
@@ -149,6 +151,7 @@ class ProjectSerializer(
         )
 
         exclude_from_list_view = [
+            "settings",
             "input_root",
             "output_root",
             "extra_data",
@@ -403,7 +406,7 @@ def get_model_serializer(model_class):
 def get_serializer_fields(model_class):
     """
     Return a list of fields declared on the Serializer that are related to the
-    a given `model_class`.
+    given `model_class`.
     """
     serializer = get_model_serializer(model_class)
     fields = list(serializer().get_fields().keys())