+++ scanpipe/pipelines/scan_codebase.py
@@ -22,7 +22,6 @@
 
 from scanpipe import pipes
 from scanpipe.pipelines import Pipeline
-from scanpipe.pipes import flag
 from scanpipe.pipes import scancode
 from scanpipe.pipes.input import copy_inputs
 
@@ -46,13 +45,11 @@ class ScanCodebase(Pipeline):
             cls.copy_inputs_to_codebase_directory,
             cls.extract_archives,
             cls.collect_and_create_codebase_resources,
-            cls.tag_empty_files,
+            cls.flag_empty_files,
+            cls.flag_ignored_resources,
             cls.scan_for_application_packages,
             cls.scan_for_files,
         )
-
-    # Set to True to extract recursively nested archives in archives.
-    extract_recursively = False
 
     def copy_inputs_to_codebase_directory(self):
         """
@@ -65,7 +62,7 @@ class ScanCodebase(Pipeline):
         """Extract archives with extractcode."""
         extract_errors = scancode.extract_archives(
             location=self.project.codebase_path,
-            recurse=self.extract_recursively,
+            recurse=self.env.get("extract_recursively", True),
         )
 
         if extract_errors:
@@ -78,10 +75,6 @@ class ScanCodebase(Pipeline):
                 project=self.project,
                 location=str(resource_path),
             )
-
-    def tag_empty_files(self):
-        """Flag empty files."""
-        flag.flag_empty_codebase_resources(self.project)
 
     def scan_for_application_packages(self):
         """Scan unknown resources for packages information."""